Allow fast-find of compendium item IDs
Looping over the compendium index and `await`ing `.getDocument` for each document in large compendiums can create a noticeable delay on drag-drop. Instead, try to look up the item ID in the compendium's index before resorting to a full search.

Note that this is a "lowest hanging fruit" change, and further performance enhancements may still be possible.